dev-requirements.txt
autopep8==1.5.4
pycodestyle==2.6.0
pylint==2.6.0
python-Levenshtein==0.12.0
setuptools==50.3.0
wheel==0.35.1
twine==3.2.0